==> Multivariados/4. Análisis Discriminante y Modelos Mixtos.Rmd <==
---
title: "reto2"
author: "Miranda Isabel Rada Chau - A01285243"
date: "`r Sys.Date()`"
output:
  word_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Carga de datos
```{r}
#Cargamos la base de datos limpia a un dataframe. 
df<- read.csv("/Users/mirisrc/Downloads/df_v5.csv")
head(df,40)
```

# Modificación de la variable date. 

```{r, warning=FALSE}
#Cambiamos el formato de las fechas a un formato Date
library(base)
library(lubridate)

df$date<- as.Date(df$date,"%d/%m/%y")


```


```{r}
#Impresión del df con las fechas transformadas. 
head(df,5)
```

```{r}
library(lubridate) 
library(dplyr)


# Función para determinar la estación
determinar_estacion <- function(fecha) {
  mes <- month(fecha)
  estacion <- case_when(
    mes %in% c(3, 4, 5) ~ 1, #Primavera
    mes %in% c(6, 7, 8) ~ 2, # Verano
    mes %in% c(9, 10, 11) ~ 3,#Otoño
    TRUE ~ 4 # Invierno
  )
  return(estacion)
}

# Agregar la columna de estaciones al DataFrame
df <- df %>%
  mutate(season = determinar_estacion(date))

# Mostrar el DataFrame resultante
head(df)
```

# Creación de periodos


```{r}
library(dplyr)

# Asumimos que df$date es una columna con fechas
df <- df %>%
  mutate(
    periodo = case_when(
      date <= "2020-03-16" ~ 1,
      date > "2020-03-16" & date <= "2021-03-16" ~ 2,
      date > "2021-03-16" & date <= "2022-03-16" ~ 3,
      date > "2022-03-16" & date <= "2023-03-16" ~ 4,
      TRUE ~ 5
    )
  )

# Mostrar los primeros y últimos valores de df$periodo
head(df, 5)
tail(df, 5)

```

```{r}
library(dplyr)

# Determinar si el periodo pertenece o no a la pandemia.
df <- df %>%
  mutate(
    pandemia = case_when(
      periodo == 2 | periodo == 3 ~ 0,
      TRUE ~ 1
    )
  )

# Mostrar los primeros y últimos valores de df. 
head(df, 5)
tail(df, 5)

```

# Creación de subsets

```{r}
#subset de antes de la pandemia 
em20<- subset(df,subset=date<="2020-03-16")

tail(em20,5)
```

```{r}
# Año 0, año de la pandemia
a0<- subset(df,subset=(date>"2020-03-16" & date<="2021-03-16"))
tail(a0,5)
```
```{r}
# Porcentaje de NOX en el año 0
sum(is.na(a0$NOX))/nrow(a0)*100

# Porcentaje de NO2 en el año 0
sum(is.na(a0$NO2))/nrow(a0)*100

# Porcentaje de NO en el año 0
sum(is.na(a0$NO))/nrow(a0)*100
```

```{r}
#Subset de 1 año después de la pandemia 
a1<- subset(df,subset=(date>"2021-03-16" & date<="2022-03-16"))
tail(a1,5)

```

```{r}
#Subset de 2 años después de la pandemia 
a2<- subset(df,subset=(date>"2022-03-16" & date<="2023-03-16"))
tail(a2,5)
```

```{r}
# #Subset de 3 años después de la pandemia (incluye todo el año 2024)
a3<- subset(df,subset=(date>"2020-03-16"))
tail(a3,5)
```
# Creación de columna binaria

### Pandemia

- a0
- a1

## No pandemia

- em20
- a2
- a3


```{r}
#Determina si el periodo corresponde a la pandemia. 
df$pandemia <- factor(ifelse(df$periodo %in% c(2, 3), "pandemia", "no pandemia"))
head(df$pandemia)
```

```{r}
#Elimina los valores nulos del df. 
df_sinNA <- na.omit(df)
```

```{r}
#Se hace un muestreo del df de 10,000 datos. 
muestra <- df_sinNA[sample(nrow(df_sinNA), 10000), ]
muestra<-muestra[ , c(-1,-17)]
head(muestra)
```
```{r}
#Se calculan las medias de todas las columnas de con respecto a la variable de la pandemia. 
medias <- aggregate(. ~ pandemia, data = muestra, mean)
medias
```

```{r}
# Probabilidades a priori
prob <- prop.table(table(muestra$pandemia))
prob
```
### Análisis discriminante lineal
```{r}
#Creación del modelo lda. 
library(MASS)
modelo <- lda(pandemia ~ CO + NO + NO2+NOX+O3+PM10+PM2.5+PRS+RAINF+RH+SO2+SR+TOUT+WSR+WDR, data = muestra)
modelo
```

```{r}
#Genera predicciones con el lda y las modela. 
library(ggplot2)
predicciones <- predict(modelo)$x
muestra$pred <- predicciones[,1]

ggplot(muestra, aes(x = pred, fill = pandemia)) +
  geom_histogram(binwidth = 0.5, alpha = 0.8, position = "identity") +
  labs(title = "Histograma de Valores Discriminantes", x = "pred", y = "Frecuencia")
```

```{r}
# Segmentación predicha por el modelo
pred_2 <- predict(modelo)$class
muestra$Prediccion <- pred_2

ggplot(muestra, aes(x = CO, y = periodo, color = Prediccion)) +
  geom_point() +
  labs(title = "Segmentación Predicha por el Modelo", x = "CO", y = "periodo")
```


```{r}
#Creación de la matriz de confianza con los valores reales y las predicciones. 
conf_matrix <- table(muestra$pandemia, muestra$Prediccion)
conf_matrix
precision <- sum(diag(conf_matrix)) / sum(conf_matrix)
error <- 1 - precision

cat("La precisión del modelo es", precision, "\n")
```
## Verificación de supuestos
```{r}
#Prueba de normalidad multivariada. 
library(MVN)
mvn(muestra[, c("CO", "NO", "NO2","NOX","O3","PM10","PM2.5","PRS","RAINF","RH","SO2","SR","TOUT","WSR","WDR")])
```

```{r}
# Verificación de homocedasticidad
library(heplots)

boxM(muestra[,c("CO", "NO", "NO2","NOX","O3","PM10","PM2.5","PRS","RAINF","RH","SO2","SR","TOUT","WSR","WDR","periodo")], muestra$pandemia)
```

# Creación de los modelos mixtos para cada variable

```{r}
#Gráfico de correlaciones para todas las variables númericas. 
library(corrplot)

correlacion <- cor(df_sinNA[c(-1,-17, -18, -19 , -20)])
corrplot(correlacion, method = "circle")

```
```{r}
#Generar una función para ordenar las correlaciones y solo imprimir las mayores a 0.3. 
library(dplyr)

correlacion_df <- as.data.frame(as.table(correlacion))

correlacion_ordenada <- correlacion_df %>%
  filter(Var1 != Var2) %>% 
  arrange(desc(abs(Freq)))

# Suponiendo que ya tenemos el DataFrame correlacion_ordenada
correlaciones_fuertes <- correlacion_ordenada %>%
  filter(abs(Freq) >= 0.3)

correlaciones_fuertes
```
#### CO

```{r}
# Filtrar las correlaciones donde CO está involucrada
correlaciones_con_CO <- correlaciones_fuertes %>%
  filter(Var1 == "CO")

correlaciones_con_CO
```

```{r}
#Creación del modelo mixto para CO. 
library(lme4)
library(Matrix)
mod_CO <- lmer(CO ~ NOX+PM2.5 +season+ (1 | pandemia)+ (1 | WSR)+ (1 | WDR), data = df_sinNA)
summary(mod_CO)
```

```{r}
#Distribución de los puntos de CO con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
library(ggplot2)
ggplot(df_sinNA, aes(x = date, y = CO)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante CO a lo largo del tiempo", x = "fecha", y = "CO") +
  theme_minimal()
```
#### Verificación de los supuestos. 
```{r}
#verificación de normalidad de la pandemia 

qqnorm(ranef(mod_CO)$pandemia$'(Intercept)', main = "Q-Q Plot de Efectos Aleatorios: Intercepto")
qqline(ranef(mod_CO)$pandemia$'(Intercept)')
```

```{r}
#residuales
qqnorm(resid(mod_CO), main = "Q-Q Plot de Residuos")
qqline(resid(mod_CO))

```

```{r}
#homocedasticidad
ggplot(df_sinNA, aes(x = fitted(mod_CO), y = resid(mod_CO))) +
  geom_point() +
  labs(title = "Residuos vs. Predicciones",
       x = "Predicciones",
       y = "Residuos") +
  theme_minimal()



```
```{r}
#independencia
library(lmtest)
dwtest(resid(mod_CO) ~ fitted(mod_CO))

```
```{r}
#linealidad con respecto a las variables
ggplot(df_sinNA, aes(x = date, y = resid(mod_CO))) +
  geom_point() +
  labs(title = "Residuos vs. Fecha",
       x = "Fecha",
       y = "Residuos") +
  theme_minimal()

```

#### NO

```{r}
# Filtrar las correlaciones donde NO está involucrada
correlaciones_con_NO <- correlaciones_fuertes %>%
  filter(Var1 == "NO")

correlaciones_con_NO
```

```{r}
#Creación del modelo mixto para NO
library(lme4)
library(Matrix)
mod_NO <- lmer(NO ~ NO2+NOX+PM10+PM2.5+season + (1 | pandemia )+(1 | WSR)+(1 | WDR ), data = df_sinNA)
summary(mod_NO)
```
```{r}
#Distribución de los puntos de NO con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
ggplot(df_sinNA, aes(x = date, y = NO)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante NO a lo largo del tiempo", x = "fecha", y = "NO") +
  theme_minimal()
```

#### NO2

```{r}
# Filtrar las correlaciones donde NO2 está involucrada
correlaciones_con_NO2 <- correlaciones_fuertes %>%
  filter(Var1 == "NO2")

correlaciones_con_NO2
```

```{r}
#Creación del modelo mixto para NO2
library(lme4)
library(Matrix)
mod_NO2 <- lmer(NO2 ~ NO+NOX+PM10+PM2.5+season + (1 | pandemia)+(1 | WSR)+(1 | WDR ), data = df_sinNA)
summary(mod_NO2)
```
```{r}
#Distribución de los puntos de NO2 con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
ggplot(df_sinNA, aes(x = date, y = NO2)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante NO2 a lo largo del tiempo", x = "fecha", y = "NO2") +
  theme_minimal()
```
#### NOX

```{r}
# Filtrar las correlaciones donde NOX está involucrada
correlaciones_con_NOX <- correlaciones_fuertes %>%
  filter(Var1 == "NOX")

correlaciones_con_NOX
```

```{r}
#Creación del modelo mixto para NOX
library(lme4)
library(Matrix)
mod_NOX <- lmer(NOX ~ CO + NO2+NO+O3+PM10+PM2.5+season + (1 | pandemia)+(1 | WSR)+(1 | WDR ), data = df_sinNA)
summary(mod_NOX)
```
```{r}
#Distribución de los puntos de NOX con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
ggplot(df_sinNA, aes(x = date, y = NOX)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante NOX a lo largo del tiempo", x = "fecha", y = "NOX") +
  theme_minimal()
```
#### O3

```{r}
# Filtrar las correlaciones donde O3 está involucrada
correlaciones_con_O3 <- correlaciones_fuertes %>%
  filter(Var1 == "O3")

correlaciones_con_O3
```

```{r}
#Creación del modelo mixto para O3
library(lme4)
library(Matrix)
mod_O3 <- lmer(O3 ~ NOX+RH+SR+TOUT+season + (1 | pandemia) + (1 | WSR) +(1 | WDR), data = df_sinNA)
summary(mod_O3)
```
```{r}
#Distribución de los puntos de O3 con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
ggplot(df_sinNA, aes(x = date, y = O3)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante O3 a lo largo del tiempo", x = "fecha", y = "O3") +
  theme_minimal()
```
#### PM10

```{r}
# Filtrar las correlaciones donde PM10 está involucrada
correlaciones_con_PM10 <- correlaciones_fuertes %>%
  filter(Var1 == "PM10")

correlaciones_con_PM10
```

```{r}
#Creación del modelo mixto para PM10
library(lme4)
library(Matrix)
mod_PM10 <- lmer(PM10 ~ NO2+NOX+NO+PM2.5+season + (1 | pandemia)+ (1 | WSR) +(1 | WDR), data = df_sinNA)
summary(mod_PM10)
```
```{r}
#Distribución de los puntos de PM10 con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
ggplot(df_sinNA, aes(x = date, y = PM10)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante PM10 a lo largo del tiempo", x = "fecha", y = "PM10") +
  theme_minimal()
```
#### Verificación de los supuestos. 
```{r}
#verificación de normalidad

qqnorm(ranef(mod_PM10)$pandemia$'(Intercept)', main = "Q-Q Plot de Efectos Aleatorios: Intercepto")
qqline(ranef(mod_PM10)$pandemia$'(Intercept)')



#residuales
qqnorm(resid(mod_PM10), main = "Q-Q Plot de Residuos")
qqline(resid(mod_PM10))

ggplot(df_sinNA, aes(x = resid(mod_PM10))) +
  geom_histogram(binwidth = 1) +
  labs(title = "Histograma de Residuos",
       x = "Residuo",
       y = "Frecuencia")

#homocedasticidad
ggplot(df_sinNA, aes(x = fitted(mod_PM10), y = resid(mod_PM10))) +
  geom_point() +
  labs(title = "Residuos vs. Predicciones",
       x = "Predicciones",
       y = "Residuos") +
  theme_minimal()

#independencia
library(lmtest)
dwtest(resid(mod_PM10) ~ fitted(mod_PM10))


#linealidad con respecto a las variables
ggplot(df_sinNA, aes(x = date, y = resid(mod_PM10))) +
  geom_point() +
  labs(title = "Residuos vs. Fecha",
       x = "Fecha",
       y = "Residuos") +
  theme_minimal()
```

#### PM2.5

```{r}
# Filtrar las correlaciones donde PM2.5 está involucrada
correlaciones_con_PM2.5 <- correlaciones_fuertes %>%
  filter(Var1 == "PM2.5")

correlaciones_con_PM2.5
```

```{r}
#Creación del modelo mixto para PM2.5
library(lme4)
library(Matrix)
mod_PM2.5 <- lmer(PM2.5 ~ CO + NO2+NOX+PM10+NO+season + (1 | pandemia)+ (1 | WSR) +(1 | WDR), data = df_sinNA)
summary(mod_PM2.5)
```
```{r}
#Distribución de los puntos de PM2.5 con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
ggplot(df_sinNA, aes(x = date, y = PM2.5)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante PM2.5 a lo largo del tiempo", x = "fecha", y = "PM2.5") +
  theme_minimal()
```
#### Verificación de los supuestos. 
```{r}
#verificación de normalidad

qqnorm(ranef(mod_PM2.5)$pandemia$'(Intercept)', main = "Q-Q Plot de Efectos Aleatorios: Intercepto")
qqline(ranef(mod_PM2.5)$pandemia$'(Intercept)')



#residuales
qqnorm(resid(mod_PM2.5), main = "Q-Q Plot de Residuos")
qqline(resid(mod_PM2.5))

ggplot(df_sinNA, aes(x = resid(mod_PM2.5))) +
  geom_histogram(binwidth = 1) +
  labs(title = "Histograma de Residuos",
       x = "Residuo",
       y = "Frecuencia")

#homocedasticidad
ggplot(df_sinNA, aes(x = fitted(mod_PM2.5), y = resid(mod_PM2.5))) +
  geom_point() +
  labs(title = "Residuos vs. Predicciones",
       x = "Predicciones",
       y = "Residuos") +
  theme_minimal()

#independencia
library(lmtest)
dwtest(resid(mod_PM2.5) ~ fitted(mod_PM2.5))


#linealidad con respecto a las variables
ggplot(df_sinNA, aes(x = date, y = resid(mod_PM2.5))) +
  geom_point() +
  labs(title = "Residuos vs. Fecha",
       x = "Fecha",
       y = "Residuos") +
  theme_minimal()
```
#### SO2

```{r}
# Filtrar las correlaciones donde SO2 está involucrada
correlaciones_con_SO2 <- correlaciones_fuertes %>%
  filter(Var1 == "SO2")

correlaciones_con_SO2
```

```{r}
#Creación del modelo mixto para SO2
library(lme4)
library(Matrix)
mod_SO2 <- lmer(SO2 ~ O3+season+ (1 | pandemia)+ (1 | WSR) +(1 | WDR), data = df_sinNA)
summary(mod_SO2)
```
```{r}
#Distribución de los puntos de SO2 con respecto a la fecha tomando en cuenta la pandemia y la estación del año. 
ggplot(df_sinNA, aes(x = date, y = SO2)) +
  geom_point(aes(color=factor(pandemia),shape=factor(season))) +
  labs(title = "Contaminante SO2 a lo largo del tiempo", x = "fecha", y = "SO2") +
  theme_minimal()
```
#### Verificación de los supuestos. 
```{r}
#verificación de normalidad

qqnorm(ranef(mod_SO2)$pandemia$'(Intercept)', main = "Q-Q Plot de Efectos Aleatorios: Intercepto")
qqline(ranef(mod_SO2)$pandemia$'(Intercept)')



#residuales
qqnorm(resid(mod_SO2), main = "Q-Q Plot de Residuos")
qqline(resid(mod_SO2))

ggplot(df_sinNA, aes(x = resid(mod_SO2))) +
  geom_histogram(binwidth = 1) +
  labs(title = "Histograma de Residuos",
       x = "Residuo",
       y = "Frecuencia")

#homocedasticidad
ggplot(df_sinNA, aes(x = fitted(mod_SO2), y = resid(mod_SO2))) +
  geom_point() +
  labs(title = "Residuos vs. Predicciones",
       x = "Predicciones",
       y = "Residuos") +
  theme_minimal()

#independencia
library(lmtest)
dwtest(resid(mod_SO2) ~ fitted(mod_SO2))


#linealidad con respecto a las variables
ggplot(df_sinNA, aes(x = date, y = resid(mod_SO2))) +
  geom_point() +
  labs(title = "Residuos vs. Fecha",
       x = "Fecha",
       y = "Residuos") +
  theme_minimal()
```
```{r}
cat("\033[3mEl modelo generado para CO es\033[0m: \n" )
cat("\n")
mod_CO
cat("\n")
cat("\033[3mEl modelo generado para NO es\033[0m: \n" )
cat("\n")
mod_NO
cat("\n")
cat("\033[3mEl modelo generado para NO2 es\033[0m: \n" )
cat("\n")
mod_NO2
cat("\n")
cat("\033[3mEl modelo generado para NOX es\033[0m: \n" )
cat("\n")
mod_NOX
cat("\n")
cat("\033[3mEl modelo generado para O3 es\033[0m: \n" )
cat("\n")
mod_O3
cat("\n")
cat("\033[3mEl modelo generado para PM10 es\033[0m: \n" )
cat("\n")
mod_PM10
cat("\n")
cat("\033[3mEl modelo generado para PM2.5 es\033[0m: \n" )
cat("\n")
mod_PM2.5
cat("\n")
cat("\033[3mEl modelo generado para SO2 es\033[0m: \n" )
cat("\n")
mod_SO2
```


## Analisis discriminante cuadratico 
```{r}
df_sinNA
```


```{r}
#Creación de la muestra para el modelo qda. 
set.seed(10)
sample <- sample(c(TRUE, FALSE), nrow(df_sinNA[c(-1,-17, -18)]), replace=TRUE, prob=c(0.8,0.2))
train<- df_sinNA[c(-1,-17,-18)][sample,]
test<- df_sinNA[c(-1,-17,-18)][!sample,]
```

```{r}
#Creación del modelo qda. 
library(MASS)
modelo_q <- qda(pandemia ~., data = train)
modelo_q
```
```{r}
#Generar predicciones según el modelo qda y la muestra de prueba. 
predicciones<- predict(modelo_q, test)
names(predicciones)
```
# Evalua el modelo 
```{r}
#precisión del modelo
mean(predicciones$class==test$pandemia)
```
# Normalidad de los residuos
```{r}
residuos<- test$pandemia != predicciones$class
library(nortest)
lillie.test(residuos)
```
```{r}
#Matriz de confusión 
table(predicciones$class,test$pandemia)
```
